server/requirements.txt
fastapi==0.115.0
uvicorn==0.34.0
pydantic==2.7.1
playwright==1.41.1
beautifulsoup4==4.13.0
requests==2.32.0
python-multipart==0.0.9
aiofiles==23.2.1
PyJWT==2.8.0
passlib==1.7.4
bcrypt==3.2.2
email-validator==2.1.0
pdfkit==1.0.0
wcag-contrast-ratio==0.9
pytest==7.4.4
pytest-asyncio==0.23.5
black==23.12.1
flake8==7.0.0
google-generativeai==0.3.2
